third_party/WebKit/LayoutTests/fast/spatial-navigation/snav-offscreen-content-expected.txt

offscreen content
Right link

Down link

PASS successfullyParsed is true

TEST COMPLETE

PASS gFocusedDocument.activeElement.getAttribute("id") is "1"
